+++ system.properties
@@ -1,1 +1,1 @@
-java.runtime.version=17
+java.runtime.version=11
